refactor: Move config class to configger.py
- Remove duplication of Config class in `creator.py` and `server.py`
- D R Y
- Also now creator generates `html.jinja` templates instead of plain `.html`
